Updated support for patch method
Currently the PATCH method is used to update the value of the leaf in the YANG model, but the same
behaviour is obtained using the PUT method. This last statement is wrong, because the PUT should be
used to replace the current resource and not to update the leaf value.
This requires some modifications in the template that will come with the next commits.