Implemented a first draft of validation
This commits add the possibility to validate the json object and in particular if the fields specified as mandatory in the YANG model are set
in the body of the request.
There still a problem when the mandatory field is a key because in this case the CLI passes the field in the path and not in the body.
In this case we cannot perform a validation.