Add support for YANG actions
This commit adds support for YANG actions.
Every action is converted into a method in the corresponding class and the
input and output models are converted into specific object available under the
serializer folder.

This commit adds two methods to the DefaultGenerator. The first method is used
to skip some files from the generation of the model. For example actions objects should not be
used to create new classes that the user modifies but they are only used as serializer/deserializer objects.
The second methods adds the possibility to specify multiple folders for the models depending on its name. This should simplify the
creation of different folders that in the current version is done manually.

Finally, some minor changes and bug fixes have been implemented.